feat: create api service for groq
