+++ app/apis/groq/openai/v1/resources/base.rb
@@ -1,0 +1,21 @@
+module Groq
+  module Openai
+    module V1
+      module Resources
+        class Base
+          attr_reader :client
+
+          def initialize(client)
+            @client = client
+          end
+
+          private
+
+          def request(**args)
+            client.request(**args)
+          end
+        end
+      end
+    end
+  end
+end
